Packaged everything come dio comanda
- Ported bloodhoundanalytics.py to Python3 using 2to3
- DBCreator now works thanks to the fixes from @voutilad
- Added .devcontainer to quickly setup dev environment in vscode
- Added Dockerfile
- exposed DBCreator and bloodhoundanalytics tools as two scripts in the
  package (bh-dbcreator and db-analytics)
- Made DBCreator automatically run the do_clear_and_generate functioon
  if the NEO4J_URL / NEO4J_USERNAME / NEO4J_PASSWORD env variables are
passed
