+++ README.md
@@ -5,3 +5,7 @@ This is a collection of miscellaneous tools released by the BloodHound team. See
 * DBCreator - Tool to generate randomized Neo4j databases for use with BloodHound
 * BloodHoundAnalytics.pbix - Proof of concept charting capability
 * BloodHoundAnalytics.py - Proof of concept audit script
+
+# Credits
+
+This package was created with [Cookiecutter](https://github.com/cookiecutter/cookiecutter) and the [byt3bl33d3r/pythoncookie](https://github.com/byt3bl33d3r/pythoncookie) project template.
