+++ bh_tools/db_creator.py
@@ -1,4 +1,3 @@
-# Requirements - pip install neo4j-driver
 # This script is used to create randomized sample databases.
 # Commands
 # 	dbconfig - Set the credentials and URL for the database you're connecting too
@@ -10,6 +9,7 @@
 #	clear_and_generate - Connects to the database, clears the DB, sets the schema, and generates random data
 
 from neo4j import GraphDatabase
+import pkgutil
 import cmd
 import os
 import sys
@@ -53,11 +53,13 @@ class MainMenu(cmd.Cmd):
         self.domain = "TESTLAB.LOCAL"
         self.current_time = int(time.time())
         self.base_sid = "S-1-5-21-883232822-274137685-4173207997"
-        with open('first.pkl', 'rb') as f:
-            self.first_names = pickle.load(f)
 
-        with open('last.pkl', 'rb') as f:
-            self.last_names = pickle.load(f)
+        self.first_names = pickle.loads(
+            pkgutil.get_data(__name__, 'data/first.pkl')
+        )
+        self.last_names = pickle.loads(
+            pkgutil.get_data(__name__, 'data/last.pkl')
+        )
 
         cmd.Cmd.__init__(self)
 
@@ -170,18 +172,14 @@ class MainMenu(cmd.Cmd):
 
         print("Resetting Schema")
         for constraint in session.run("CALL db.constraints"):
-            session.run("DROP {}".format(constraint['description']))
+            session.run("DROP CONSTRAINT {}".format(constraint['name']))
 
         for index in session.run("CALL db.indexes"):
-            session.run("DROP {}".format(index['description']))
-
-        session.run(
-            "CREATE CONSTRAINT id_constraint ON (c:Base) ASSERT c.objectid IS UNIQUE")
-        session.run("CREATE INDEX name_index FOR (n:Base) ON (n.name)")
+            session.run("DROP INDEX {}".format(index['name']))
 
         session.close()
 
-        print("DB Cleared and Schema Set")
+        print("DB Cleared and Schema Reset")
 
     def test_db_conn(self):
         self.connected = False
@@ -201,7 +199,7 @@ class MainMenu(cmd.Cmd):
 
     def do_clear_and_generate(self, args):
         self.test_db_conn()
-        self.do_cleardb("a")
+        self.do_cleardb(args)
         self.generate_data()
 
     def split_seq(self, iterable, size):
@@ -250,6 +248,11 @@ class MainMenu(cmd.Cmd):
 
         print("Starting data generation with nodes={}".format(self.num_nodes))
 
+        print("Generating Schema")
+        session.run(
+            "CREATE CONSTRAINT id_constraint ON (c:Base) ASSERT c.objectid IS UNIQUE")
+        session.run("CREATE INDEX name_index FOR (n:Base) ON (n.name)")
+
         print("Populating Standard Nodes")
         base_statement = "MERGE (n:Base {name: $gname}) SET n:Group, n.objectid=$sid"
         session.run(f"{base_statement},n.highvalue=true",
@@ -274,7 +277,7 @@ class MainMenu(cmd.Cmd):
         base_statement = "MERGE (n:Base {name:$gpo, objectid:$guid}) SET n:GPO"
         session.run(base_statement, gpo=cn("DEFAULT DOMAIN POLICY"), guid=ddp)
         session.run(base_statement, gpo=cn(
-            "DEFAULT DOMAIN CONTROLLERS POLICY"), guid=ddp)
+            "DEFAULT DOMAIN CONTROLLERS POLICY"), guid=ddcp)
         session.run("MERGE (n:Base {name:$ou, objectid:$guid, blocksInheritance: false}) SET n:OU", ou=cn(
             "DOMAIN CONTROLLERS"), guid=dcou)
 
@@ -353,10 +356,10 @@ class MainMenu(cmd.Cmd):
 
             if (len(props) > 500):
                 session.run(
-                    'UNWIND {props} as prop MERGE (n:Base {objectid: prop.id}) SET n:Computer, n += prop.props WITH n MERGE (m:Group {name:$gname}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props, gname=group_name)
+                    'UNWIND $props as prop MERGE (n:Base {objectid: prop.id}) SET n:Computer, n += prop.props WITH n MERGE (m:Group {name:$gname}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props, gname=group_name)
                 props = []
         session.run(
-            'UNWIND {props} as prop MERGE (n:Base {objectid:prop.id}) SET n:Computer, n += prop.props WITH n MERGE (m:Group {name:$gname}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props, gname=group_name)
+            'UNWIND $props as prop MERGE (n:Base {objectid:prop.id}) SET n:Computer, n += prop.props WITH n MERGE (m:Group {name:$gname}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props, gname=group_name)
 
         print("Creating Domain Controllers")
         for state in states:
@@ -401,11 +404,11 @@ class MainMenu(cmd.Cmd):
             }})
             if (len(props) > 500):
                 session.run(
-                    'UNWIND {props} as prop MERGE (n:Base {objectid:prop.id}) SET n:User, n += prop.props WITH n MATCH (m:Group {name:$gname}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props, gname=group_name)
+                    'UNWIND $props as prop MERGE (n:Base {objectid:prop.id}) SET n:User, n += prop.props WITH n MATCH (m:Group {name:$gname}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props, gname=group_name)
                 props = []
 
         session.run(
-            'UNWIND {props} as prop MERGE (n:Base {objectid:prop.id}) SET n:User, n += prop.props WITH n MATCH (m:Group {name:$gname}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props, gname=group_name)
+            'UNWIND $props as prop MERGE (n:Base {objectid:prop.id}) SET n:User, n += prop.props WITH n MATCH (m:Group {name:$gname}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props, gname=group_name)
 
         print("Generating Group Nodes")
         weighted_parts = ["IT"] * 7 + ["HR"] * 13 + \
@@ -420,11 +423,11 @@ class MainMenu(cmd.Cmd):
             props.append({'name': group_name, 'id': sid})
             if len(props) > 500:
                 session.run(
-                    'UNWIND {props} as prop MERGE (n:Base {objectid:prop.id}) SET n:Group, n.name=prop.name', props=props)
+                    'UNWIND $props as prop MERGE (n:Base {objectid:prop.id}) SET n:Group, n.name=prop.name', props=props)
                 props = []
 
         session.run(
-            'UNWIND {props} as prop MERGE (n:Base {objectid:prop.id}) SET n:Group, n.name=prop.name', props=props)
+            'UNWIND $props as prop MERGE (n:Base {objectid:prop.id}) SET n:Group, n.name=prop.name', props=props)
 
         print("Adding Domain Admins to Local Admins of Computers")
         session.run(
@@ -458,11 +461,11 @@ class MainMenu(cmd.Cmd):
 
             if (len(props) > 500):
                 session.run(
-                    'UNWIND {props} AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Group {name:prop.b}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props)
+                    'UNWIND $props AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Group {name:prop.b}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props)
                 props = []
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Group {name:prop.b}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props)
+            'UNWIND $props AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Group {name:prop.b}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props)
 
         print("Adding users to groups")
         props = []
@@ -496,11 +499,11 @@ class MainMenu(cmd.Cmd):
 
             if len(props) > 500:
                 session.run(
-                    'UNWIND {props} AS prop MERGE (n:User {name:prop.a}) WITH n,prop MERGE (m:Group {name:prop.b}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props)
+                    'UNWIND $props AS prop MERGE (n:User {name:prop.a}) WITH n,prop MERGE (m:Group {name:prop.b}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props)
                 props = []
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:User {name:prop.a}) WITH n,prop MERGE (m:Group {name:prop.b}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props)
+            'UNWIND $props AS prop MERGE (n:User {name:prop.a}) WITH n,prop MERGE (m:Group {name:prop.b}) WITH n,m MERGE (n)-[:MemberOf]->(m)', props=props)
 
         it_users = it_users + das
         it_users = list(set(it_users))
@@ -534,7 +537,7 @@ class MainMenu(cmd.Cmd):
 
             if len(props) > 500:
                 session.run(
-                    'UNWIND {props} AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (n)-[:AdminTo]->(m)', props=props)
+                    'UNWIND $props AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (n)-[:AdminTo]->(m)', props=props)
                 props = []
 
         for x in super_groups:
@@ -543,11 +546,11 @@ class MainMenu(cmd.Cmd):
 
             if len(props) > 500:
                 session.run(
-                    'UNWIND {props} AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (n)-[:AdminTo]->(m)', props=props)
+                    'UNWIND $props AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (n)-[:AdminTo]->(m)', props=props)
                 props = []
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (n)-[:AdminTo]->(m)', props=props)
+            'UNWIND $props AS prop MERGE (n:Group {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (n)-[:AdminTo]->(m)', props=props)
 
         print("Adding RDP/ExecuteDCOM/AllowedToDelegateTo")
         count = int(math.floor(len(computers) * .1))
@@ -558,7 +561,7 @@ class MainMenu(cmd.Cmd):
             props.append({'a': user, 'b': comp})
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:User {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:CanRDP]->(m)', props=props)
+            'UNWIND $props AS prop MERGE (n:User {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:CanRDP]->(m)', props=props)
 
         props = []
         for i in range(0, count):
@@ -567,7 +570,7 @@ class MainMenu(cmd.Cmd):
             props.append({'a': user, 'b': comp})
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:User {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:ExecuteDCOM]->(m)', props=props)
+            'UNWIND $props AS prop MERGE (n:User {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:ExecuteDCOM]->(m)', props=props)
 
         props = []
         for i in range(0, count):
@@ -576,7 +579,7 @@ class MainMenu(cmd.Cmd):
             props.append({'a': user, 'b': comp})
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:Group {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:CanRDP]->(m)', props=props)
+            'UNWIND $props AS prop MERGE (n:Group {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:CanRDP]->(m)', props=props)
 
         props = []
         for i in range(0, count):
@@ -585,7 +588,7 @@ class MainMenu(cmd.Cmd):
             props.append({'a': user, 'b': comp})
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:Group {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:ExecuteDCOM]->(m)', props=props)
+            'UNWIND $props AS prop MERGE (n:Group {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:ExecuteDCOM]->(m)', props=props)
 
         props = []
         for i in range(0, count):
@@ -594,7 +597,7 @@ class MainMenu(cmd.Cmd):
             props.append({'a': user, 'b': comp})
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:User {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:AllowedToDelegate]->(m)', props=props)
+            'UNWIND $props AS prop MERGE (n:User {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:AllowedToDelegate]->(m)', props=props)
 
         props = []
         for i in range(0, count):
@@ -605,7 +608,7 @@ class MainMenu(cmd.Cmd):
             props.append({'a': user, 'b': comp})
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:Computer {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:AllowedToDelegate]->(m)', props=props)
+            'UNWIND $props AS prop MERGE (n:Computer {name: prop.a}) MERGE (m:Computer {name: prop.b}) MERGE (n)-[r:AllowedToDelegate]->(m)', props=props)
 
         print("Adding sessions")
         max_sessions_per_user = int(math.ceil(math.log10(self.num_nodes)))
@@ -624,11 +627,11 @@ class MainMenu(cmd.Cmd):
 
             if (len(props) > 500):
                 session.run(
-                    'UNWIND {props} AS prop MERGE (n:User {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (m)-[:HasSession]->(n)', props=props)
+                    'UNWIND $props AS prop MERGE (n:User {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (m)-[:HasSession]->(n)', props=props)
                 props = []
 
         session.run(
-            'UNWIND {props} AS prop MERGE (n:User {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (m)-[:HasSession]->(n)', props=props)
+            'UNWIND $props AS prop MERGE (n:User {name:prop.a}) WITH n,prop MERGE (m:Computer {name:prop.b}) WITH n,m MERGE (m)-[:HasSession]->(n)', props=props)
 
         print("Adding Domain Admin ACEs")
         group_name = "DOMAIN ADMINS@{}".format(self.domain)
@@ -638,33 +641,33 @@ class MainMenu(cmd.Cmd):
 
             if len(props) > 500:
                 session.run(
-                    'UNWIND {props} as prop MATCH (n:Computer {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
+                    'UNWIND $props as prop MATCH (n:Computer {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
                 props = []
 
         session.run(
-            'UNWIND {props} as prop MATCH (n:Computer {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
+            'UNWIND $props as prop MATCH (n:Computer {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
 
         for x in users:
             props.append({'name': x})
 
             if len(props) > 500:
                 session.run(
-                    'UNWIND {props} as prop MATCH (n:User {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
+                    'UNWIND $props as prop MATCH (n:User {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
                 props = []
 
         session.run(
-            'UNWIND {props} as prop MATCH (n:User {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
+            'UNWIND $props as prop MATCH (n:User {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
 
         for x in groups:
             props.append({'name': x})
 
             if len(props) > 500:
                 session.run(
-                    'UNWIND {props} as prop MATCH (n:Group {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
+                    'UNWIND $props as prop MATCH (n:Group {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
                 props = []
 
         session.run(
-            'UNWIND {props} as prop MATCH (n:Group {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
+            'UNWIND $props as prop MATCH (n:Group {name:prop.name}) WITH n MATCH (m:Group {name:$gname}) WITH m,n MERGE (m)-[r:GenericAll {isacl:true}]->(n)', props=props, gname=group_name)
 
         print("Creating OUs")
         temp_comps = computers
@@ -682,11 +685,11 @@ class MainMenu(cmd.Cmd):
                 props.append({'compname': c, 'ouguid': guid, 'ouname': ouname})
                 if len(props) > 500:
                     session.run(
-                        'UNWIND {props} as prop MERGE (n:Computer {name:prop.compname}) WITH n,prop MERGE (m:Base {objectid:prop.ouguid, name:prop.ouname, blocksInheritance: false}) SET m:OU WITH n,m,prop MERGE (m)-[:Contains]->(n)', props=props)
+                        'UNWIND $props as prop MERGE (n:Computer {name:prop.compname}) WITH n,prop MERGE (m:Base {objectid:prop.ouguid, name:prop.ouname, blocksInheritance: false}) SET m:OU WITH n,m,prop MERGE (m)-[:Contains]->(n)', props=props)
                     props = []
 
         session.run(
-            'UNWIND {props} as prop MERGE (n:Computer {name:prop.compname}) WITH n,prop MERGE (m:Base {objectid:prop.ouguid, name:prop.ouname, blocksInheritance: false}) SET m:OU WITH n,m,prop MERGE (m)-[:Contains]->(n)', props=props)
+            'UNWIND $props as prop MERGE (n:Computer {name:prop.compname}) WITH n,prop MERGE (m:Base {objectid:prop.ouguid, name:prop.ouname, blocksInheritance: false}) SET m:OU WITH n,m,prop MERGE (m)-[:Contains]->(n)', props=props)
 
         temp_users = users
         random.shuffle(temp_users)
@@ -703,11 +706,11 @@ class MainMenu(cmd.Cmd):
                 props.append({'username': c, 'ouguid': guid, 'ouname': ouname})
                 if len(props) > 500:
                     session.run(
-                        'UNWIND {props} as prop MERGE (n:User {name:prop.username}) WITH n,prop MERGE (m:Base {objectid:prop.ouguid, name:prop.ouname, blocksInheritance: false}) SET m:OU WITH n,m,prop MERGE (m)-[:Contains]->(n)', props=props)
+                        'UNWIND $props as prop MERGE (n:User {name:prop.username}) WITH n,prop MERGE (m:Base {objectid:prop.ouguid, name:prop.ouname, blocksInheritance: false}) SET m:OU WITH n,m,prop MERGE (m)-[:Contains]->(n)', props=props)
                     props = []
 
         session.run(
-            'UNWIND {props} as prop MERGE (n:User {name:prop.username}) WITH n,prop MERGE (m:Base {objectid:prop.ouguid, name:prop.ouname, blocksInheritance: false}) SET m:OU WITH n,m,prop MERGE (m)-[:Contains]->(n)', props=props)
+            'UNWIND $props as prop MERGE (n:User {name:prop.username}) WITH n,prop MERGE (m:Base {objectid:prop.ouguid, name:prop.ouname, blocksInheritance: false}) SET m:OU WITH n,m,prop MERGE (m)-[:Contains]->(n)', props=props)
 
         props = []
         for x in list(ou_guid_map.keys()):
@@ -715,7 +718,7 @@ class MainMenu(cmd.Cmd):
             props.append({'b': guid})
 
         session.run(
-            'UNWIND {props} as prop MERGE (n:OU {objectid:prop.b}) WITH n MERGE (m:Domain {name:$domain}) WITH n,m MERGE (m)-[:Contains]->(n)', props=props, domain=self.domain)
+            'UNWIND $props as prop MERGE (n:OU {objectid:prop.b}) WITH n MERGE (m:Domain {name:$domain}) WITH n,m MERGE (m)-[:Contains]->(n)', props=props, domain=self.domain)
 
         print("Creating GPOs")
 
@@ -762,35 +765,35 @@ class MainMenu(cmd.Cmd):
                 p = random.choice(all_principals)
                 p2 = random.choice(gpos)
                 session.run(
-                    'MERGE (n:Group {name:{group}}) MERGE (m {name:{principal}}) MERGE (n)-' + ace_string + '->(m)', group=i, principal=p)
-                session.run('MERGE (n:Group {name:{group}}) MERGE (m:GPO {name:{principal}}) MERGE (n)-' +
+                    'MERGE (n:Group {name:$group}) MERGE (m {name:$principal}) MERGE (n)-' + ace_string + '->(m)', group=i, principal=p)
+                session.run('MERGE (n:Group {name:$group}) MERGE (m:GPO {name:$principal}) MERGE (n)-' +
                             ace_string + '->(m)', group=i, principal=p2)
             elif ace == 'AddMember':
                 p = random.choice(it_groups)
                 session.run(
-                    'MERGE (n:Group {name:{group}}) MERGE (m:Group {name:{principal}}) MERGE (n)-' + ace_string + '->(m)', group=i, principal=p)
+                    'MERGE (n:Group {name:$group}) MERGE (m:Group {name:$principal}) MERGE (n)-' + ace_string + '->(m)', group=i, principal=p)
             elif ace == 'ReadLAPSPassword':
                 p = random.choice(all_principals)
                 targ = random.choice(computers)
                 session.run(
-                    'MERGE (n {name:{principal}}) MERGE (m:Computer {name:{target}}) MERGE (n)-[r:ReadLAPSPassword]->(m)', target=targ, principal=p)
+                    'MERGE (n {name:$principal}) MERGE (m:Computer {name:$target}) MERGE (n)-[r:ReadLAPSPassword]->(m)', target=targ, principal=p)
             else:
                 p = random.choice(it_users)
                 session.run(
-                    'MERGE (n:Group {name:{group}}) MERGE (m:User {name:{principal}}) MERGE (n)-' + ace_string + '->(m)', group=i, principal=p)
+                    'MERGE (n:Group {name:$group}) MERGE (m:User {name:$principal}) MERGE (n)-' + ace_string + '->(m)', group=i, principal=p)
 
         print("Marking some users as Kerberoastable")
         i = random.randint(10, 20)
         i = min(i, len(it_users))
         for user in random.sample(it_users, i):
             session.run(
-                'MATCH (n:User {name:{user}}) SET n.hasspn=true', user=user)
+                'MATCH (n:User {name:$user}) SET n.hasspn=true', user=user)
 
         print("Adding unconstrained delegation to a few computers")
         i = random.randint(10, 20)
         i = min(i, len(computers))
         session.run(
-            'MATCH (n:Computer {name:{user}}) SET n.unconstrainteddelegation=true', user=user)
+            'MATCH (n:Computer {name:$user}) SET n.unconstrainteddelegation=true', user=user)
 
         session.run('MATCH (n:User) SET n.owned=false')
         session.run('MATCH (n:Computer) SET n.owned=false')
@@ -800,10 +803,26 @@ class MainMenu(cmd.Cmd):
 
         print("Database Generation Finished!")
 
+def main():
+    menu = MainMenu()
+    NEO4J_URL = os.environ.get("NEO4J_URL")
+    NEO4J_USERNAME = os.environ.get("NEO4J_USERNAME")
+    NEO4J_PASSWORD = os.environ.get("NEO4J_PASSWORD")
 
-if __name__ == '__main__':
-    try:
-        MainMenu().cmdloop()
-    except KeyboardInterrupt:
-        print("Exiting")
-        sys.exit()
+    NEO4J_ENCRYPTION = False
+    if os.environ.get("NEO4J_ENCRYPTION"):
+        NEO4J_ENCRYPTION = bool(int(os.environ["NEO4J_ENCRYPTION"]))
+
+    if NEO4J_URL and NEO4J_USERNAME and NEO4J_PASSWORD:
+        menu.url = NEO4J_URL
+        menu.username = NEO4J_USERNAME
+        menu.password = NEO4J_PASSWORD
+        menu.use_encryption = NEO4J_ENCRYPTION
+        menu.do_clear_and_generate(None)
+
+    else:
+        try:
+            menu.cmdloop()
+        except KeyboardInterrupt:
+            print("Exiting")
+            sys.exit()
